Improved distortion map performance
git-svn-id: http://yuvtoolkit.googlecode.com/svn@148 fe7d2ec1-bd83-e687-9167-72ef5a783f6b

+++ trunk/Plugins/YT_MeasuresBasic/YT_MeasuresBasic.h
@@ -7,7 +7,7 @@ class MeasuresBasic : public QObject, public Measure
 {
 	Q_OBJECT;
 
-	double ComputeMSE(FramePtr input1, FramePtr input2, int plane, DistMapPtr mseMap);
+	float ComputeMSE(FramePtr input1, FramePtr input2, int plane, DistMapPtr mseMap);
 
 	MeasureCapabilities m_Capabilities;
 public:
